fix: restructure demand folder to match scan structure
