+++ beam/tests/setup.py
@@ -1,9 +1,7 @@
 import datetime
-import types
 from itertools import groupby
 
 import frappe
-from erpnext.accounts.doctype.account.account import update_account_number
 from erpnext.manufacturing.doctype.production_plan.production_plan import (
 	get_items_for_material_requests,
 )
@@ -11,7 +9,7 @@ from erpnext.setup.utils import enable_all_roles_and_domains, set_defaults_for_t
 from erpnext.stock.get_item_details import get_item_details
 from frappe.desk.page.setup_wizard.setup_wizard import setup_complete
 
-from beam.beam.demand import build_demand_map
+from beam.beam.demand.demand import build_demand_map
 from beam.tests.fixtures import boms, customers, items, operations, suppliers, workstations
